Added Home/Menu Controller
Added a Home controller which will help control  the Home, about, and Gallery page since these will be more like a static website. There is little to no code for these views and classes. I added a Menu Controller with will handle the Menu Page View and the Add/Delete/Create controlls for the Menu Items.