Added Menu Pages
Added a Menu PAge for the User to view and then create a create/edit page for admin use only. Also the delete is also for admin use only. You can click  login at the bottom of the Menu page and this will take you to login . The tables show grouped Menu Items.